make: Restrict install-cli to the current working directory
Commit 5840834fa887e19ff328a038b38b756b3c474da0 introduced
copying of scripts and escripts to the plugin currently
being worked on. At some point after that, in recent weeks,
the target would run for all applications being compiled,
and not just the plugin we have asked Make to build. This
led to increased build times.

This change ensures the scripts and escripts are only
copied for the directory we have ran Make in. So if
we simply do "make" this is the top-level directory.
If we do "make -C deps/rabbit" this is the rabbit
application's directory.

Doing this has a huge impact on performance when
rebuilding with no changes done to the source code:

  make  6,63s user 2,60s system 106% cpu 8,668 total
  make  2,88s user 1,30s system 117% cpu 3,567 total

And a less pronounced impact on specific applications
(results will vary):

  make -C deps/rabbit  0,50s user 0,18s system 100% cpu 0,677 total
  make -C deps/rabbit  0,43s user 0,12s system 101% cpu 0,551 total